task-08/pyproject.toml
[project]
name = "task-08"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "lyricsgenius>=3.0.1",
    "wordcloud>=1.9.3",
    "matplotlib>=3.8.0",
    "streamlit>=1.30.0",
]
